[Game] fixed ActionSlots; [Game] fixed Doodads spawn; [Packets] fixed SCUnitStatePacket; [Data] updated transfer_path.xml, doodad_spawns.json, npc_spawns.json, gimmick_spawns.json, transfer_spawns.json; [Game] disable SendSwitchZoneState; [Game] disable movement NPC in AiManager;

+++ AAEmu.Game/Core/Packets/C2G/CSMoveUnitPacket.cs
@@ -131,7 +131,7 @@ namespace AAEmu.Game.Core.Packets.C2G
                     return;
                 }
 
-                if ((mType.ActorFlags & 0x20) != 0)
+                if (((ushort)mType.ActorFlags & 0x20) != 0)
                 {
                     Connection
                     .ActiveChar
